seperate concerns for caught exceptions
