--- td_biblio/exceptions.py
@@ -1,15 +1,0 @@
-# td_biblio exceptions
-class BaseLoaderError(Exception):
-    pass
-
-
-class BibTeXLoaderError(BaseLoaderError):
-    pass
-
-
-class PMIDLoaderError(BaseLoaderError):
-    pass
-
-
-class DOILoaderError(BaseLoaderError):
-    pass
