+++ td_biblio/exceptions/bibtex.py
@@ -1,0 +1,5 @@
+from td_biblio.exceptions.base import BaseLoaderError
+
+
+class BibTeXLoaderError(BaseLoaderError):
+    pass
