+++ td_biblio/exceptions/doi.py
@@ -1,0 +1,5 @@
+from td_biblio.exceptions.base import BaseLoaderError
+
+
+class DOILoaderError(BaseLoaderError):
+    pass
