+++ td_biblio/exceptions/pmid.py
@@ -1,0 +1,5 @@
+from td_biblio.exceptions.base import BaseLoaderError
+
+
+class PMIDLoaderError(BaseLoaderError):
+    pass
