+++ td_biblio/factories/__init__.py
@@ -7,8 +7,11 @@ from .abstract_entity import * # noqa
 from .abstract_human import * # noqa
 from .author import * # noqa
 from .author_entry_rank import * # noqa
-from .entry import * # noqa
-from .editor import * # noqa
 from .collection import * # noqa
+from .editor import * # noqa
+from .entry import * # noqa
+from .entry_with_author import * # noqa
+from .entry_with_static_author import * # noqa
+from .fuzzy_pages_attribute import * # noqa
 from .journal import * # noqa
 from .publisher import * # noqa
