+++ td_biblio/factories/abstract_entity.py
@@ -1,0 +1,8 @@
+from factory.django import DjangoModelFactory
+from td_biblio.models import AbstractEntity
+
+
+class AbstractEntityFactory(DjangoModelFactory):
+    class Meta:
+        model = AbstractEntity
+        abstract = True
